Fix #1247: Adapt getClass for primitives to Scala.js subtyping
Note that the checkfile override for run/t6318_primitives.check for
Scala 2.11.0 and 2.11.1 has been wrongly introduced in
efe48f93214f376d99e6409fe94acba0af45d620. This commit both fixes the
mistake and adapts them for the change.

+++ test-suite/src/test/scala/scala/scalajs/testsuite/compiler/ReflectionTest.scala
@@ -44,12 +44,12 @@ object ReflectionTest extends JasmineTest {
         c.asInstanceOf[js.Any]
       expect((false: Any).getClass).toBe(classOf[java.lang.Boolean])
       expect(('a': Any).getClass).toBe(classOf[java.lang.Character])
-      expect((1.toByte: Any).getClass).toBe(classOf[java.lang.Integer])
-      expect((1.toShort: Any).getClass).toBe(classOf[java.lang.Integer])
-      expect((1: Any).getClass).toBe(classOf[java.lang.Integer])
+      expect((1.toByte: Any).getClass).toBe(classOf[java.lang.Byte])
+      expect((1.toShort: Any).getClass).toBe(classOf[java.lang.Byte])
+      expect((1: Any).getClass).toBe(classOf[java.lang.Byte])
       expect((1L: Any).getClass).toBe(classOf[java.lang.Long])
-      expect((1.5f: Any).getClass).toBe(classOf[java.lang.Double])
-      expect((1.5: Any).getClass).toBe(classOf[java.lang.Double])
+      expect((1.5f: Any).getClass).toBe(classOf[java.lang.Float])
+      expect((1.5: Any).getClass).toBe(classOf[java.lang.Float])
       expect(((): Any).getClass).toBe(classOf[scala.runtime.BoxedUnit])
     }
